requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
transformers>=4.25.0

# Data processing
numpy>=1.24.0
pandas>=1.5.0
opencv-python>=4.7.0
Pillow>=9.4.0

# Roboflow integration
roboflow>=1.0.0

# Evaluation
datasets>=2.10.0
evaluate>=0.4.0

# Visualization
matplotlib>=3.6.0
tqdm>=4.65.0

# Logging
tensorboard>=2.12.0

# Optional but recommended
scipy>=1.10.0
scikit-learn>=1.2.0